The logic for creating or updating the validation file has been moved to ValidationUtil. Implemented processing of unknown components
