+++ plugins/ru.runa.gpd.form.ftl/src/ru/runa/gpd/formeditor/wysiwyg/FormEditor.java
@@ -72,6 +72,7 @@ import ru.runa.gpd.ui.view.SelectionProvider;
 import ru.runa.gpd.util.EditorUtils;
 import ru.runa.gpd.util.IOUtils;
 import ru.runa.gpd.util.VariableUtils;
+import ru.runa.gpd.validation.ValidationUtil;
 import ru.runa.wfe.InternalApplicationException;
 
 import com.google.common.base.Objects;
@@ -150,7 +151,7 @@ public class FormEditor extends MultiPageEditorPart implements IResourceChangeLi
             public void propertyChanged(Object source, int propId) {
                 if (propId == FormEditor.CLOSED) {
                     if (formFile.exists()) {
-                        EditorUtils.createOrUpdateValidation(formNode, formFile);
+                        ValidationUtil.createOrUpdateValidation(formNode, formFile);
                     }
                     boolean formEditorsAvailable = false;
                     IWorkbenchPage workbenchPage = PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage();
@@ -380,7 +381,7 @@ public class FormEditor extends MultiPageEditorPart implements IResourceChangeLi
         sourceEditor.doSave(monitor);
         if (formNode != null) {
             formNode.setDirty();
-            EditorUtils.createOrUpdateValidation(formNode, formFile);
+            ValidationUtil.createOrUpdateValidation(formNode, formFile);
         }
         if (isBrowserLoaded()) {
             browser.execute("setHTMLSaved()");
